Fix the hung issue in 2c011ce
We have two ways to notify the QE senders to stop provide tuples.
1) send a stop message through interconnect 2) set QueryFinishPending
flag through libpq. Method 1 can not resolve the hung issue in 2c011ce
if the test case in that PR also contains a rescan RESULT node.

For method 2, current problem is once a QE sender stuck at waiting
interconnect data acks because out of sending buffers or stuck at
waiting EOS acks, the QE sender has no chance to process
QueryFinishPending flag. The fix break the loops and set the
stillActive to false in a rude way, so adding more assertion and
filter for safe.