Fixes wrong result with correlated queries (#6074), the issue
This commit fixes wrong results with correlated queries (#6074),
the issue is introduced by commit 2c011ce in which it made a bad
decision to call ExecSquelchNode() in RESULT node.

The original thought in 2c011ce is, once the one-time filter qual
of a RESULT node is evaluated to be false, we no longer need to
fetch tuple from child node, so it's safe to send a stop message
to source QE senders. Then if the RESULT node is rescanned,
one-time filter is revaluated to be true, RESULT node need to fetch
tuple from child node, however, the QE senders has been stopped and
some tuples are missed. So the first step is reverting the changes
in 2c011ce and then add a new test case.