Enable master auto-failover
This is based on the implementation of primary/mirror segment failover.

An FTS process is started on one of the primary segment, naming 'master
prober'. This process will probe master periodically, once it finds the master
not responding, it will mark the master down and promote the standby.

When the master finds the standby is out-of-sync, all the transactions will be
blocked until 'master prober' is noticed in the next probe.

FTS process manage the 'master prober', it watches the state of 'master
prober'; starts 'master prober' when it's down; restart it when the
master/standby information is changed (for example, a new standby is added by
gpinitstandby); start 'master prober' on another segment if the old segment is
down.

see the discussion:
https://groups.google.com/a/greenplum.org/forum/#!msg/gpdb-dev/e9F5rvUQuWA/qsZfdUFSAgAJ
see the design and progress:
https://github.com/greenplum-db/gpdb/wiki/Master-auto-failover